Extract files github.py, redirects.py, settings.py
